pipelines/scout/etl/requirements.txt
# Scout ETL Pipeline Dependencies
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Database & API clients
supabase>=2.0.0
sqlalchemy>=2.0.0
pyodbc>=4.0.39
requests>=2.31.0

# File handling
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Utilities
blake3>=0.3.3
tqdm>=4.66.0

# Development & testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0